updated controllers to split out new listings from other routes
