--- app/controllers/api/v1/newlistings_controller.rb
@@ -1,10 +1,0 @@
-# NOTE: This file is not currenntly being used
-# This is the next stage and in progress
-
-module Api
-  module V1
-    class NewListingsController < ApplicationController
-
-    end
-  end
-end
